+++ js/main.js
@@ -1,2 +1,2 @@
 import './create-rentors.js';
-import './generation-text.js';
+import './adding-an-ad.js';
